requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
bandit==1.7.0
billiard==3.6.3.0
black==20.8b1
celery==5.0.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-stubs==1.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djangorestframework-stubs==1.3.0
flake8==3.8.4
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
importlib-metadata==3.1.1
iniconfig==1.1.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
pbr==5.5.1
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.25.0
six==1.15.0
smmap==3.0.4
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
google-auth==1.24.0
django-filter==2.4.0
drf-yasg==1.20.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
